requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.28
pymongo==4.5.0
dvc==3.21.1
dvc-s3==2.23.0
joblib==1.3.1
python-dotenv==1.0.0
mlflow==2.8.0
tqdm==4.66.1 